Organize Swift classes/structs/extensions
