+++ lib/cocoa/src/TEnum.swift
@@ -1,0 +1,20 @@
+//
+//  TEnum.swift
+//  Pods
+//
+//  Created by Kevin Wooten on 10/7/15.
+//
+//
+
+import Foundation
+
+
+public protocol TEnum : TSerializable {
+  
+}
+
+public extension TEnum {
+  
+  public static var thriftType : TType { return TType.I32 }
+  
+}
